[Webpack] implement differential serving
