+++ .browserslistrc
@@ -1,0 +1,11 @@
+[legacy]
+Last 2 versions
+IE >= 11
+not IE 10
+not ie_mob 10
+iOS >= 8
+
+[modern]
+chrome >= 80
+firefox >= 80
+safari >= 12
