+++ .gitignore
@@ -35,6 +35,8 @@ web/app/themes/wwp_child_theme/assets/final/*.gif
 web/app/themes/wwp_child_theme/assets/final/*.png
 web/app/themes/wwp_child_theme/assets/final/*.jpg
 web/app/themes/wwp_child_theme/assets/final/*.jpeg
+web/app/themes/wwp_child_theme/assets/final/legacy
+web/app/themes/wwp_child_theme/assets/final/modern
 
 # Documentation
 web/documentation/dev/dest
