+++ config/webpack/webpack.legacy.config.js
@@ -1,0 +1,6 @@
+const {getCustomConfig} = require('./webpack.base.config');
+const {getAssetsEntries, BUILD_TYPE} = require('./paths');
+
+const config = getCustomConfig(BUILD_TYPE.LEGACY, getAssetsEntries(BUILD_TYPE.LEGACY));
+
+module.exports = config;
