+++ package.json
@@ -4,10 +4,9 @@
   "description": "wonderwp base poc",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack",
-    "build:prod": "webpack --env=production",
-    "start": "webpack-dev-server --open",
-    "watch": "webpack --watch",
+    "watch": "encore dev --watch -c ./config/webpack/webpack.modern.config.js & encore dev --watch -c ./config/webpack/webpack.legacy.config.js",
+    "build:prod": "encore production --progress -c config/webpack/webpack.modern.config.js & encore production --progress -c config/webpack/webpack.legacy.config.js & wait",
+    "build": "encore dev --progress -c config/webpack/webpack.modern.config.js & encore dev --progress -c config/webpack/webpack.legacy.config.js & wait",
     "sprites": "node compile-sprites.js"
   },
   "repository": {
@@ -21,37 +20,14 @@
   },
   "homepage": "https://github.com/wdfproje/wonderwp-base#readme",
   "devDependencies": {
-    "autoprefixer": "^8.5.0",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015-ie": "^6.7.0",
-    "clean-webpack-plugin": "^0.1.17",
-    "copy-webpack-plugin": "^4.5.1",
-    "css-loader": "^2.1.1",
-    "cssnano": "^4.1.10",
-    "expose-loader": "^0.7.3",
-    "file-loader": "^1.1.11",
-    "file-system": "^2.2.2",
-    "html-webpack-plugin": "^2.30.1",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.14.1",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "postcss-loader": "^2.1.5",
-    "rebuild-node-sass": "^1.1.0",
-    "resolve-url-loader": "^3.1.2",
-    "sass-loader": "^6.0.6",
-    "style-loader": "0.19.0",
+    "@symfony/webpack-encore": "^1.5.0",
+    "node-sass": "^6.0.1",
+    "sass-loader": "^12.1.0",
     "svg-sprite": "^1.3.7",
     "vinyl": "^2.1.0",
-    "webpack": "^4.39.1",
-    "webpack-bundle-analyzer": "^3.4.1",
-    "webpack-cli": "^3.3.6",
-    "webpack-version-file": "^0.1.3",
-    "webpackbar": "^1.5.1"
+    "webpack-notifier": "^1.13.0"
   },
   "dependencies": {
-    "autoprefixer": "^8.4.1",
     "bxslider": "^4.2.13",
     "jquery": "^3.5.1",
     "jquery-lazy": "^1.7.7",
@@ -60,7 +36,6 @@
     "modaal": "^0.3.1",
     "pewjs": "^1.0.4",
     "pikaday": "^1.7.0",
-    "sass-resources-loader": "^1.3.4",
     "scrollmagic": "github:janpaepke/ScrollMagic#master",
     "selectric": "^1.13.0",
     "svg4everybody": "^2.1.9"
