--- postcss.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-    plugins: {
-        autoprefixer: {
-            browsers: ['last 2 versions', 'iOS >= 8'],
-            grid: true
-        }
-    }
-};
-
-/*
-* See https://github.com/postcss/autoprefixer
-* and https://github.com/browserslist/browserslist#queries
-* for finer browser autoprefixer configuration.
-* */
