+++ web/app/themes/wwp_child_theme/assets/raw/js/app_bootstrap.js
@@ -1,66 +1,6 @@
 import {EventManager} from "./features/EventManager";
 import {FeatureDetector} from "./features/FeatureDetector";
 
-/**
- * Object.assign Polyfill
- */
-if (typeof Object.assign != 'function') {
-  Object.assign = function (target) {
-    'use strict';
-    if (target == null) {
-      throw new TypeError('Cannot convert undefined or null to object');
-    }
-
-    target = Object(target);
-    for (var index = 1; index < arguments.length; index++) {
-      var source = arguments[index];
-      if (source != null) {
-        for (var key in source) {
-          if (Object.prototype.hasOwnProperty.call(source, key)) {
-            target[key] = source[key];
-          }
-        }
-      }
-    }
-    return target;
-  };
-}
-//Object.entries polyfill
-if (!Object.entries) {
-  Object.entries = function( obj ){
-    var ownProps = Object.keys( obj ),
-      i = ownProps.length,
-      resArray = new Array(i); // preallocate the Array
-    while (i--)
-      resArray[i] = [ownProps[i], obj[ownProps[i]]];
-
-    return resArray;
-  };
-}
-//Str.includes polyfill
-if (!String.prototype.includes) {
-  String.prototype.includes = function(search, start) {
-    if (typeof start !== 'number') {
-      start = 0;
-    }
-
-    if (start + search.length > this.length) {
-      return false;
-    } else {
-      return this.indexOf(search, start) !== -1;
-    }
-  };
-}
-
-String.prototype.replaceArray = function(find, replace) {
-  var replaceString = this;
-  var regex;
-  for (var i = 0; i < find.length; i++) {
-    regex = new RegExp(find[i], "g");
-    replaceString = replaceString.replace(regex, replace[i]);
-  }
-  return replaceString;
-};
 
 /**
  * Pew.js polyfill
