+++ web/app/themes/wwp_child_theme/assets/raw/scss/theme.scss
@@ -9,7 +9,7 @@
 @import "menus/megamenu/megamenu";
 @import "menus/barre-menu-mobile/barre-menu-mobile";
 
-@import '/../../../includes/Components/components';
+@import '../../../includes/Components/components';
 
 @import "layout";
 @import "utilities";
